+++ src/ops_data_store/cli/db/__init__.py
@@ -32,6 +32,60 @@ def check() -> None:
         raise typer.Abort() from e
 
 
+@app.command(help="Configure a new DB for use.")
+def setup() -> None:
+    """
+    Create required Postgres extensions and functions.
+
+    It's expected this command will be run once as part of setting up a new database as part of an instance of this
+    application. It MUST be assumed however that this command may be run multiple times (possibly as part of ongoing
+    automation) and so this command MUST NOT put existing data at risk or return errors for pre-existing objects etc.
+
+    Error handling is intentionally omitted in this command as the underlying exceptions will give useful context
+    should any command fail, and wrapping these exceptions won't provide any benefit.
+    """
+    logger.info("Setting up required database objects.")
+    print("Setting up database for first time use.")
+    print(
+        "Note: If this command fails, please either create an issue in the 'Ops Data Store' project in GitLab, or "
+        "contact MAGIC at magic@bas.ac.uk with the output of this command."
+    )
+
+    with psycopg.connect(config.DB_DSN) as conn, conn.cursor() as cur:
+        for extension in ["postgis", "pgcrypto"]:
+            logger.info(f"Setting up required DB extension {extension}.")
+            # exempting SQL injection check as extension names are effectively fixed
+            cur.execute(f"CREATE EXTENSION IF NOT EXISTS {extension}")
+            cur.execute(f"SELECT COUNT(name) FROM pg_available_extensions WHERE name='{extension}'")  # noqa: S608
+            if cur.fetchone()[0] != 1:  # pragma: no cover - see MAGIC/ops-data-store#43
+                logger.error(f"Required extension '{extension}' not found after creating.")
+                print(f"No. Required extension '{extension}' not found.")
+                raise typer.Abort()
+            logger.info(f"Required DB extension '{extension}' ok.")
+
+        logger.info("Setting up required DB function 'generate_ulid'.")
+        cur.execute(
+            """
+        CREATE OR REPLACE FUNCTION generate_ulid() RETURNS uuid
+            AS $$
+                SELECT (
+                    lpad(to_hex(floor(extract(epoch FROM clock_timestamp()) * 1000)::bigint), 12, '0') ||
+                    encode(gen_random_bytes(10), 'hex')
+                )::uuid;
+            $$ LANGUAGE SQL;
+        """
+        )
+        cur.execute("""SELECT COUNT(*) FROM pg_proc WHERE proname = 'generate_ulid';""")
+        if cur.fetchone()[0] != 1:  # pragma: no cover - see MAGIC/ops-data-store#43
+            logger.error("Required function 'generate_ulid' not found after creating.")
+            print("No. Required function 'generate_ulid' not found.")
+            raise typer.Abort()
+        logger.info("Required DB function 'generate_ulid' ok.")
+
+    logger.info("Database setup complete.")
+    print("Complete.")
+
+
 @app.command(help="Execute contents of an SQL against DB.")
 def run(input_path: Annotated[Path, typer.Option()]) -> None:
     """Load contents of SQL file and execute against DB."""