+++ tests/ops_data_store_tests/cli/test_cli_db.py
@@ -48,6 +48,7 @@ class TestCliDBSetup:
 
         assert "Required DB extension 'postgis' ok." in caplog.text
         assert "Required DB extension 'pgcrypto' ok." in caplog.text
+        assert "Required DB extension 'fuzzystrmatch' ok." in caplog.text
         assert "Required DB function 'generate_ulid' ok." in caplog.text
         assert "Required DB function 'geom_as_ddm' ok." in caplog.text
         assert "Required DB data type 'ddm_point' ok." in caplog.text