+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 
 * `db setup` CLI command
   [#33](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/33)
+* Postgres function for formatting coordinates in DDM format
+  [#46](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/46)
 
 ## [0.2.0] - 2023-10-04