+++ CHANGELOG.md
@@ -15,11 +15,17 @@ Wherever possible, a reference to an issue in the project issue tracker should b
   [#33](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/33)
 * Postgres function for formatting coordinates in DDM format
   [#46](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/46)
+* Documentation on required/expected dataset identifiers
+  [#36](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/36)
+* Placeholder schemas and seed data for depot and instrument managed datasets
+  [#34](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/34)
 
 ### Fixed
 
 * Adding requirement/checks for 'fuzzystrmatch' Postgres extension
   [#47](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/47)
+* Documenting how the dotenv module searches for `.env` files
+  [#57](https://gitlab.data.bas.ac.uk/MAGIC/ops-data-store/-/issues/57)
 
 ## [0.2.0] - 2023-10-04